+++ src/constants.ts
@@ -1,6 +1,6 @@
 /**
  * Constants and pricing tables
- * Version: 0.13.2
+ * Version: 0.13.6
  */
 
 export interface ModelPricing {